+++ sw/lib/include/neorv32_xip.h
@@ -76,6 +76,10 @@ enum NEORV32_XIP_CTRL_enum {
   XIP_CTRL_SPI_CSEN       = 21, /**< XIP control register(21) (r/w): SPI chip-select enable */
   XIP_CTRL_HIGHSPEED      = 22, /**< XIP control register(22) (r/w): SPI high-speed mode enable (ignoring XIP_CTRL_PRSC) */
   XIP_CTRL_BURST_EN       = 23, /**< XIP control register(23) (r/w): Enable XIP burst mode */
+  XIP_CTRL_CDIV0          = 24, /**< XIP control register(24) (r/w): Clock divider bit 0 */
+  XIP_CTRL_CDIV1          = 25, /**< XIP control register(25) (r/w): Clock divider bit 1 */
+  XIP_CTRL_CDIV2          = 26, /**< XIP control register(26) (r/w): Clock divider bit 2 */
+  XIP_CTRL_CDIV3          = 27, /**< XIP control register(27) (r/w): Clock divider bit 3 */
 
   XIP_CTRL_PHY_BUSY       = 30, /**< XIP control register(20) (r/-): SPI PHY is busy */
   XIP_CTRL_XIP_BUSY       = 31  /**< XIP control register(31) (r/-): XIP access in progress */
@@ -88,10 +92,11 @@ enum NEORV32_XIP_CTRL_enum {
  **************************************************************************/
 /**@{*/
 int  neorv32_xip_available(void);
-void neorv32_xip_setup(int prsc, int cpol, int cpha, uint8_t rd_cmd);
+void neorv32_xip_setup(int prsc, int cdiv, int cpol, int cpha, uint8_t rd_cmd);
 int  neorv32_xip_start(int abytes);
 void neorv32_xip_highspeed_enable(void);
 void neorv32_xip_highspeed_disable(void);
+uint32_t neorv32_xip_get_clock_speed(void);
 void neorv32_xip_burst_mode_enable(void);
 void neorv32_xip_burst_mode_disable(void);
 void neorv32_xip_spi_trans(int nbytes, uint64_t *rtx_data);